Implement __init_subclass__ for frame serialization registration
