+++ sunpy/coordinates/__init__.py
@@ -18,7 +18,6 @@ below (see `astropy.coordinates.builtin_frames`).
 
 """
 
-from . import _table_yaml  # noqa: F401
 from . import sun
 from ._transformations import _make_sunpy_graph, propagate_with_solar_surface, transform_with_sun_center
 from .ephemeris import *
